dcp-815 separate output.py
